step4_pipeline/requirements_step4.txt
# Requirements for Step 4: Pipeline Integration
# This step integrates all models from Step 3 into a complete pipeline

# Core ML libraries (inherited from Step 3)
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0

# ML frameworks (inherited from Step 3)
diffusers>=0.25.0
transformers>=4.35.0

# Additional utilities
einops>=0.6.0
accelerate>=0.20.0
safetensors>=0.3.0

# Optional: for better performance
xformers>=0.0.20  # Optional: for attention optimization
